Made a small change to invoice

+++ InventarVali.Models/Invoice.cs
@@ -15,7 +15,7 @@ namespace InventarVali.Models
         public int Id { get; set; }
         public string? InvoiceNumber { get; set; }
         public DateTime? InvoiceDate { get; set; }
-        public decimal? TotalPrice { get; set; }
+        public double? TotalPrice { get; set; }
         public double Price { get; set; }
         public string ?CardNumber { get; set; }
         public string? InvoiceUrl { get; set; }
